feat: enhance database error handling and validate package dates

+++ models/tourist_package.py
@@ -1,5 +1,5 @@
 """Tourist package value object."""
-
+import decimal
 from datetime import date
 
 
@@ -7,7 +7,7 @@ class TouristPackage:
     """Aggregate destination selections into a purchasable package."""
 
     def __init__(self, name: str, start_date: date, end_date: date,
-                 total_price: int):
+                 total_price: decimal):
         self.name = name
         self.start_date = start_date
         self.end_date = end_date
